+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes
 
 ### Unreleased
 
+* 2021-01-14 - Improvement: Add option to output the information banner without any particular Bootstrap color.
 * 2021-01-14 - Bugfix: The infobanner should not clean the entered HTML code too strictly.
 
 ### Release v3.9-r4